+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.0.0
+  * Bump dependency versions for singer-python and request. [#17](https://github.com/singer-io/tap-mailshake/pull/17)
+  * Fixed schema mismatch issues, updated schema for all existing streams.[#18](https://github.com/singer-io/tap-mailshake/pull/18)
+  * Changed replication method for team_members INCREMENTAL to FULL_TABLE and fixed sync issues.[#18](https://github.com/singer-io/tap-mailshake/pull/18)
+
+## 0.1.0
+  * Bump dependency versions for twistlock compliance [#16](https://github.com/singer-io/tap-mailshake/pull/16)
+
+## 0.0.10
+  * Refactor to fix linting errors [#14](https://github.com/singer-io/tap-mailshake/pull/14)
+
+## 0.0.9
+  * Bump dependency versions for twistlock compliance [#12](https://github.com/singer-io/tap-mailshake/pull/12)
+
 ## 0.0.8
   * Fix other usage of unencoded api key.